feat: improve adoption token related components
